Add feature for fetching git repo's
