+++ features/steps/generic_steps.py
@@ -1,5 +1,6 @@
 """Steps for features tests."""
 
+import os
 import subprocess
 
 from behave import then, when  # pylint: disable=no-name-in-module
@@ -8,18 +9,12 @@ from behave import then, when  # pylint: disable=no-name-in-module
 @when('I run "{cmd}"')
 def step_impl(context, cmd):
     """Call a command."""
-    subprocess.call(cmd.split())
+    assert subprocess.call(cmd.split()) == 0
 
 
-@then("it should generate the manifest '{name}'")
-def step_impl(context, name):
-    """Check a manifest."""
-    with open(name, "r") as manifest:
+@then("the following projects are fetched")
+def step_impl(context):
 
-        for actual, expected in zip(
-            context.text.splitlines(True), manifest.readlines()
-        ):
-
-            assert (
-                actual.strip() == expected.strip()
-            ), f"Actual {actual.strip()} != Expected {expected.strip()}"
+    for project in context.table:
+        assert os.path.exists(project["path"]), f"No project found at {project}"
+        assert os.listdir(project["path"]), f"{project} is just an empty directory!"
